% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphs.R
\name{add_edge}
\alias{add_edge}
\title{Add an edge from s to t}
\usage{
add_edge(g, s, t, ...)
}
\arguments{
\item{g}{A graph ACSet}

\item{s}{Source vertex index}

\item{t}{Target vertex index}

\item{...}{Additional attributes}
}
\description{
Add an edge from s to t
}
\examples{
g <- Graph()
add_vertices(g, 2)
add_edge(g, 1, 2)
edge_src(g, 1) # 1
edge_tgt(g, 1) # 2
}
