% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphs.R
\name{add_vertex}
\alias{add_vertex}
\title{Add a vertex, returning its ID}
\usage{
add_vertex(g, ...)
}
\arguments{
\item{g}{A graph ACSet}

\item{...}{Additional attributes}
}
\description{
Add a vertex, returning its ID
}
