% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uwd.R
\name{box_spec}
\alias{box_spec}
\title{Shorthand for creating a relation box spec}
\usage{
box_spec(name, ...)
}
\arguments{
\item{name}{Box name}

\item{...}{Junction names for the box ports}
}
\description{
Shorthand for creating a relation box spec
}
