% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catlab-package.R
\docType{package}
\name{catlab-package}
\alias{catlab}
\alias{catlab-package}
\title{catlab: Categorical Algebra for R}
\description{
Categorical algebra tools for R including graphs, finite categories, functors, ACSet transformations, wiring diagrams, and data migration. Built on the acsets package. Port of parts of the AlgebraicJulia Catlab.jl package.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/catRgory/catlab}
  \item \url{https://catrgory.github.io/catlab/}
  \item Report bugs at \url{https://github.com/catRgory/catlab/issues}
}

}
\author{
\strong{Maintainer}: Simon Frost \email{sdwfrost@users.noreply.github.com} (\href{https://orcid.org/0000-0002-5207-9879}{ORCID})

}
\keyword{internal}
