% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/limits.R
\name{coequalizer}
\alias{coequalizer}
\title{Coequalizer of two parallel ACSet transformations}
\usage{
coequalizer(f, g)
}
\arguments{
\item{f}{ACSetTransformation A → B}

\item{g}{ACSetTransformation A → B}
}
\value{
List with \code{coequalizer} (ACSet) and \code{proj} (projection morphism)
}
\description{
Given f, g: A → B, identify f(a) with g(a) for all a.
}
