% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fincat.R
\name{compose_transformations}
\alias{compose_transformations}
\title{Compose two ACSet transformations}
\usage{
compose_transformations(alpha, beta)
}
\arguments{
\item{alpha}{An \link{ACSetTransformation}}

\item{beta}{An \link{ACSetTransformation}}
}
\description{
Compose two ACSet transformations
}
\examples{
g1 <- path_graph(2)
g2 <- path_graph(3)
g3 <- path_graph(4)
alpha <- ACSetTransformation(
  components = list(V = c(1L, 2L), E = 1L),
  dom_acset = g1, codom_acset = g2
)
beta <- ACSetTransformation(
  components = list(V = c(1L, 2L, 3L), E = c(1L, 2L)),
  dom_acset = g2, codom_acset = g3
)
gamma <- compose_transformations(alpha, beta)
is_natural(gamma) # TRUE
}
