% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/structured_cospans.R
\name{cospan_apex}
\alias{cospan_apex}
\title{Extract the apex from a structured cospan}
\usage{
cospan_apex(sc)
}
\arguments{
\item{sc}{StructuredCospan}
}
\value{
ACSet
}
\description{
Extract the apex from a structured cospan
}
