% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/structured_cospans.R
\name{discrete_acset}
\alias{discrete_acset}
\title{Create a discrete (empty) ACSet with n parts in a given object}
\usage{
discrete_acset(schema, ob, n)
}
\arguments{
\item{schema}{BasicSchema}

\item{ob}{Character: which object to populate}

\item{n}{Integer: how many parts}
}
\value{
ACSet
}
\description{
The L functor: FinSet -> ACSet maps a finite set of size n to an ACSet
with n parts in the interface object and nothing else.
}
