% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fincat.R
\name{FinCat}
\alias{FinCat}
\title{Finite category (presented by an ACSet schema)}
\usage{
FinCat(schema = acsets::BasicSchema())
}
\arguments{
\item{schema}{A \link[acsets:BasicSchema]{acsets::BasicSchema} defining the category}
}
\description{
Finite category (presented by an ACSet schema)
}
