% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fincat.R
\name{FinFunctor}
\alias{FinFunctor}
\title{Functor between finite categories (schema morphism)}
\usage{
FinFunctor(ob_map = list(), hom_map = list(), dom = FinCat(), codom = FinCat())
}
\arguments{
\item{ob_map}{Named list mapping domain objects to codomain objects}

\item{hom_map}{Named list mapping domain morphisms/attributes to codomain
morphisms/attributes}

\item{dom}{Domain \link{FinCat}}

\item{codom}{Codomain \link{FinCat}}
}
\description{
Functor between finite categories (schema morphism)
}
\examples{
cat_g <- FinCat(schema = SchGraph)
F <- FinFunctor(
  ob_map = list(V = "V", E = "E"),
  hom_map = list(src = "src", tgt = "tgt"),
  dom = cat_g, codom = cat_g
)
F@ob_map$V # "V"
}
