% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/typed_acsets.R
\name{flatten_typed}
\alias{flatten_typed}
\title{Flatten a typed ACSet (forget the typing)}
\usage{
flatten_typed(tacs)
}
\arguments{
\item{tacs}{TypedACSet}
}
\value{
The underlying ACSet
}
\description{
Flatten a typed ACSet (forget the typing)
}
