% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/structured_cospans.R
\name{foot_sizes}
\alias{foot_sizes}
\title{Get the foot sizes of a structured cospan}
\usage{
foot_sizes(sc)
}
\arguments{
\item{sc}{StructuredCospan}
}
\value{
integer vector
}
\description{
Get the foot sizes of a structured cospan
}
