% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/limits.R
\name{from_initial}
\alias{from_initial}
\title{Unique morphism from initial object}
\usage{
from_initial(acs)
}
\arguments{
\item{acs}{An ACSet}
}
\value{
ACSetTransformation from initial(acs@schema)
}
\description{
Unique morphism from initial object
}
