% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphviz.R
\name{generic_to_dot}
\alias{generic_to_dot}
\title{Convert an arbitrary ACSet to DOT format}
\usage{
generic_to_dot(acs, ...)
}
\arguments{
\item{acs}{An ACSet}

\item{...}{Additional arguments (currently unused)}
}
\value{
A DOT format string
}
\description{
Convert an arbitrary ACSet to DOT format
}
