% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rewriting.R
\name{get_matches}
\alias{get_matches}
\title{Find all matches of a rule in a graph}
\usage{
get_matches(rule, graph, limit = Inf)
}
\arguments{
\item{rule}{A Rule object}

\item{graph}{Target ACSet}

\item{limit}{Maximum matches to find}
}
\value{
List of ACSetTransformations (L → graph)
}
\description{
Find all matches of a rule in a graph
}
