% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphs.R
\name{Graph}
\alias{Graph}
\title{Create a directed graph}
\usage{
Graph(...)
}
\arguments{
\item{...}{Arguments passed to the ACSet constructor}
}
\description{
Create a directed graph
}
\examples{
g <- Graph()
add_vertices(g, 3)
add_edge(g, 1, 2)
add_edge(g, 2, 3)
nv(g) # 3
ne(g) # 2
}
