% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fincat.R
\name{id_transformation}
\alias{id_transformation}
\title{Identity transformation}
\usage{
id_transformation(acs)
}
\arguments{
\item{acs}{An ACSet}
}
\description{
Identity transformation
}
\examples{
g <- path_graph(3)
id <- id_transformation(g)
is_natural(id) # TRUE
id@components$V  # c(1, 2, 3)
}
