% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fincat.R
\name{is_natural}
\alias{is_natural}
\title{Check naturality of an ACSet transformation}
\usage{
is_natural(alpha)
}
\arguments{
\item{alpha}{An \link{ACSetTransformation}}
}
\description{
Check naturality of an ACSet transformation
}
