% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/typed_acsets.R
\name{is_typed_morphism}
\alias{is_typed_morphism}
\title{Check if a morphism between typed ACSets preserves typing}
\usage{
is_typed_morphism(f, tacs1, tacs2)
}
\arguments{
\item{f}{ACSetTransformation from tacs1 to tacs2}

\item{tacs1}{TypedACSet (domain)}

\item{tacs2}{TypedACSet (codomain)}
}
\value{
logical
}
\description{
Given typed ACSets phi_1: X_1 -> T and phi_2: X_2 -> T and a morphism
f: X_1 -> X_2, check that phi_2 . f = phi_1 (types are preserved).
}
