% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/structured_cospans.R
\name{make_leg}
\alias{make_leg}
\title{Create a leg morphism from a discrete foot into an apex}
\usage{
make_leg(schema, interface_ob, indices, apex)
}
\arguments{
\item{schema}{BasicSchema}

\item{interface_ob}{Character: which object type}

\item{indices}{Integer vector: which apex parts this foot maps to}

\item{apex}{ACSet: the target}
}
\value{
ACSetTransformation
}
\description{
Create a leg morphism from a discrete foot into an apex
}
