% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/structured_cospans.R
\name{nlegs}
\alias{nlegs}
\title{Get the number of legs (feet) of a structured cospan}
\usage{
nlegs(sc)
}
\arguments{
\item{sc}{StructuredCospan}
}
\value{
integer
}
\description{
Get the number of legs (feet) of a structured cospan
}
