% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/structured_cospans.R
\name{oapply_cospans}
\alias{oapply_cospans}
\title{Compose a list of structured cospans via a UWD}
\usage{
oapply_cospans(w, cospans)
}
\arguments{
\item{w}{UWD (an ACSet on SchUWD)}

\item{cospans}{List of StructuredCospan objects (one per box)}
}
\value{
StructuredCospan with legs from outer ports
}
\description{
Given a UWD (undirected wiring diagram) and a list of structured cospans
(one per box), compose them by identifying legs at shared junctions.
}
\details{
This is the general oapply for structured cospans: it takes the coproduct
of all apexes, then uses the UWD's wiring to build a pushout that
identifies interface parts at shared junctions.
}
