% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/structured_cospans.R
\name{otimes_cospans}
\alias{otimes_cospans}
\title{Monoidal product of two structured cospans (side-by-side)}
\usage{
otimes_cospans(M, N)
}
\arguments{
\item{M}{StructuredCospan}

\item{N}{StructuredCospan}
}
\value{
StructuredCospan
}
\description{
Places M and N side by side via coproduct. No identification.
}
