% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphs.R
\name{path_graph}
\alias{path_graph}
\title{Path graph: 1 → 2 → ... → n}
\usage{
path_graph(n)
}
\arguments{
\item{n}{Number of vertices}
}
\description{
Path graph: 1 → 2 → ... → n
}
\examples{
g <- path_graph(4)
nv(g) # 4
ne(g) # 3
}
