% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphviz.R
\name{petri_to_dot}
\alias{petri_to_dot}
\title{Render a Petri net as a bipartite DOT graph}
\usage{
petri_to_dot(pn)
}
\arguments{
\item{pn}{A Petri net ACSet}
}
\description{
Species are circles, transitions are boxes. Input/output arcs
connect them. Uses DiagrammeR for rendering.
}
\examples{
pn <- acsets::ACSet(acsets::BasicSchema(
  obs = c("S", "T", "I", "O"),
  homs = list(acsets::hom("is", "I", "S"), acsets::hom("it", "I", "T"),
             acsets::hom("os", "O", "S"), acsets::hom("ot", "O", "T"))))
acsets::add_parts(pn, "S", 2)
acsets::add_parts(pn, "T", 1)
acsets::add_part(pn, "I", is = 1L, it = 1L)
acsets::add_part(pn, "I", is = 2L, it = 1L)
acsets::add_part(pn, "O", os = 1L, ot = 1L)
cat(petri_to_dot(pn))
}
