% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/limits.R
\name{pullback}
\alias{pullback}
\title{Pullback of ACSets along a cospan}
\usage{
pullback(f, g)
}
\arguments{
\item{f}{ACSetTransformation B → D}

\item{g}{ACSetTransformation C → D}
}
\value{
List with \code{pullback}, \code{proj1}, \code{proj2}
}
\description{
Given f: B → D and g: C → D, compute the pullback B ×_D C, consisting
of pairs (b, c) where f(b) = g(c) for every object type.
}
