% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/limits.R
\name{pushout}
\alias{pushout}
\title{Pushout of ACSets along two morphisms from a common apex}
\usage{
pushout(f, g)
}
\arguments{
\item{f}{ACSetTransformation A → B}

\item{g}{ACSetTransformation A → C}
}
\description{
Given f: A → B and g: A → C, compute the pushout B +_A C.
}
