% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphs.R
\docType{data}
\name{SchGraph}
\alias{SchGraph}
\title{Directed graph schema: V (vertices), E (edges), src, tgt}
\format{
An object of class \code{acsets::BasicSchema} (inherits from \code{S7_object}) of length 1.
}
\usage{
SchGraph
}
\description{
Directed graph schema: V (vertices), E (edges), src, tgt
}
\keyword{datasets}
