% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphs.R
\docType{data}
\name{SchLabelledGraph}
\alias{SchLabelledGraph}
\title{Labelled graph schema (vertex and edge labels)}
\format{
An object of class \code{acsets::BasicSchema} (inherits from \code{S7_object}) of length 1.
}
\usage{
SchLabelledGraph
}
\description{
Labelled graph schema (vertex and edge labels)
}
\keyword{datasets}
