% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphs.R
\docType{data}
\name{SchReflexiveGraph}
\alias{SchReflexiveGraph}
\title{Reflexive graph schema: adds refl : V → E}
\format{
An object of class \code{acsets::BasicSchema} (inherits from \code{S7_object}) of length 1.
}
\usage{
SchReflexiveGraph
}
\description{
Reflexive graph schema: adds refl : V → E
}
\keyword{datasets}
