% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphs.R
\docType{data}
\name{SchSymmetricGraph}
\alias{SchSymmetricGraph}
\title{Symmetric graph schema: adds inv : E → E (edge involution)}
\format{
An object of class \code{acsets::BasicSchema} (inherits from \code{S7_object}) of length 1.
}
\usage{
SchSymmetricGraph
}
\description{
Symmetric graph schema: adds inv : E → E (edge involution)
}
\keyword{datasets}
