% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uwd.R
\docType{data}
\name{SchUWD}
\alias{SchUWD}
\title{UWD Schema}
\format{
An object of class \code{acsets::BasicSchema} (inherits from \code{S7_object}) of length 1.
}
\usage{
SchUWD
}
\description{
UWD Schema
}
\keyword{datasets}
