% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphs.R
\docType{data}
\name{SchWeightedGraph}
\alias{SchWeightedGraph}
\title{Weighted graph schema}
\format{
An object of class \code{acsets::BasicSchema} (inherits from \code{S7_object}) of length 1.
}
\usage{
SchWeightedGraph
}
\description{
Weighted graph schema
}
\keyword{datasets}
