% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/structured_cospans.R
\name{structured_cospan}
\alias{structured_cospan}
\title{Create a structured cospan (open system)}
\usage{
structured_cospan(apex, legs, interface_ob)
}
\arguments{
\item{apex}{ACSet (the full system)}

\item{legs}{List of ACSetTransformations from feet into apex}

\item{interface_ob}{Character: which object type forms the interface}
}
\value{
StructuredCospan
}
\description{
Create a structured cospan (open system)
}
