% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/structured_cospans.R
\name{StructuredCospan}
\alias{StructuredCospan}
\title{Structured multicospan (open system with multiple legs/feet)}
\usage{
StructuredCospan(
  apex = acsets::ACSet(),
  legs = list(),
  interface_ob = character(0)
)
}
\arguments{
\item{apex}{The ACSet forming the cospan apex}

\item{legs}{List of \link{ACSetTransformation} objects from discrete feet to apex}

\item{interface_ob}{Name of the object type used for interfaces}
}
\description{
Represents an open ACSet: an apex with multiple legs mapping discrete
feet into the apex. Each leg identifies which parts of the apex are
"exposed" at that interface.
}
