% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/limits.R
\name{terminal}
\alias{terminal}
\title{Terminal object: exactly one part per object type}
\usage{
terminal(schema)
}
\arguments{
\item{schema}{A BasicSchema}
}
\value{
An ACSet with 1 part per object
}
\description{
Terminal object: exactly one part per object type
}
