% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphviz.R
\name{to_dot}
\alias{to_dot}
\title{Convert an ACSet to DOT format string}
\usage{
to_dot(x, ...)
}
\arguments{
\item{x}{An ACSet}

\item{...}{Additional arguments passed to format-specific methods}
}
\description{
Convert an ACSet to DOT format string
}
\examples{
g <- path_graph(3)
cat(to_dot(g))
}
