% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/limits.R
\name{to_terminal}
\alias{to_terminal}
\title{Unique morphism to terminal object}
\usage{
to_terminal(acs)
}
\arguments{
\item{acs}{An ACSet}
}
\value{
ACSetTransformation to terminal(acs@schema)
}
\description{
Unique morphism to terminal object
}
