% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/typed_acsets.R
\name{typed_acset}
\alias{typed_acset}
\title{Create a typed ACSet}
\usage{
typed_acset(acset, type_system, ...)
}
\arguments{
\item{acset}{The concrete ACSet}

\item{type_system}{The type system ACSet}

\item{...}{Named integer vectors: component mappings for the typing morphism}
}
\value{
TypedACSet
}
\description{
Create a typed ACSet
}
