% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/typed_acsets.R
\name{typed_coproduct}
\alias{typed_coproduct}
\title{Typed coproduct of typed ACSets}
\usage{
typed_coproduct(tacs1, tacs2)
}
\arguments{
\item{tacs1}{TypedACSet}

\item{tacs2}{TypedACSet}
}
\value{
TypedACSet
}
\description{
Given typed ACSets phi_1: X_1 -> T and phi_2: X_2 -> T, compute their
coproduct via pushout of the initial object. In practice this is the
disjoint union with combined typing.
}
\examples{
T <- Graph()
add_vertices(T, 2)
add_edge(T, 1, 2)
g1 <- Graph()
add_vertices(g1, 2)
add_edge(g1, 1, 2)
t1 <- typed_acset(g1, T, V = c(1L, 2L), E = 1L)
g2 <- Graph()
add_vertices(g2, 2)
add_edge(g2, 1, 2)
t2 <- typed_acset(g2, T, V = c(1L, 2L), E = 1L)
tc <- typed_coproduct(t1, t2)
nv(flatten_typed(tc)) # 4
}
