% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/typed_acsets.R
\name{TypedACSet}
\alias{TypedACSet}
\title{Typed ACSet: an ACSet with a typing morphism to a type system}
\usage{
TypedACSet(
  acset = acsets::ACSet(),
  type_system = acsets::ACSet(),
  typing = ACSetTransformation()
)
}
\arguments{
\item{acset}{The ACSet instance}

\item{type_system}{The type system ACSet}

\item{typing}{An \link{ACSetTransformation} from acset to type_system}
}
\description{
A TypedACSet pairs an ACSet X with a morphism phi: X -> T where T is the
"type system" (another ACSet on the same schema). Every element in X is
assigned a type by phi.
}
\examples{
# Type system: two vertex types, one edge type
T <- Graph()
add_vertices(T, 2)
add_edge(T, 1, 2)
# Instance typed over T
g <- Graph()
add_vertices(g, 3)
add_edge(g, 1, 3)
add_edge(g, 2, 3)
typing <- ACSetTransformation(
  components = list(V = c(1L, 1L, 2L), E = c(1L, 1L)),
  dom_acset = g, codom_acset = T
)
tg <- TypedACSet(acset = g, type_system = T, typing = typing)
flatten_typed(tg) # returns g
}
