% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uwd.R
\name{UWD}
\alias{UWD}
\alias{uwd}
\title{Undirected wiring diagrams}
\usage{
UWD(...)

uwd(outer, ...)
}
\arguments{
\item{...}{For \code{uwd()}: box specifications as named character vectors.
\code{infection = c("s", "i")} creates a box "infection" with ports connected
to junctions "s" and "i".}

\item{outer}{Character vector of outer junction names}
}
\value{
A UWD ACSet
}
\description{
\code{UWD} is the ACSet type constructor for undirected wiring diagrams.
\code{uwd()} is a convenience constructor using concise syntax.
}
