% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fincat.R
\name{ACSetTransformation}
\alias{ACSetTransformation}
\title{ACSet transformation (natural transformation between ACSets)}
\usage{
ACSetTransformation(
  components = list(),
  dom_acset = acsets::ACSet(),
  codom_acset = acsets::ACSet()
)
}
\arguments{
\item{components}{Named list of integer vectors mapping parts}

\item{dom_acset}{Domain ACSet}

\item{codom_acset}{Codomain ACSet}
}
\description{
ACSet transformation (natural transformation between ACSets)
}
