% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphs.R
\name{add_vertices}
\alias{add_vertices}
\title{Add multiple vertices}
\usage{
add_vertices(g, n, ...)
}
\arguments{
\item{g}{A graph ACSet}

\item{n}{Number of vertices to add}

\item{...}{Additional attributes}
}
\description{
Add multiple vertices
}
