% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fincat.R
\name{compose_transformations}
\alias{compose_transformations}
\title{Compose two ACSet transformations}
\usage{
compose_transformations(alpha, beta)
}
\arguments{
\item{alpha}{An \link{ACSetTransformation}}

\item{beta}{An \link{ACSetTransformation}}
}
\description{
Compose two ACSet transformations
}
