% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/limits.R
\name{coproduct}
\alias{coproduct}
\title{Coproduct (disjoint union) with injection morphisms}
\usage{
coproduct(acs1, acs2)
}
\arguments{
\item{acs1}{First ACSet}

\item{acs2}{Second ACSet}
}
\description{
Coproduct (disjoint union) with injection morphisms
}
\examples{
g1 <- path_graph(2) # 1 -> 2
g2 <- path_graph(2) # 1 -> 2
cp <- coproduct(g1, g2)
nv(cp$coproduct) # 4
ne(cp$coproduct) # 2
}
