% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/limits.R
\name{equalizer}
\alias{equalizer}
\title{Equalizer of two parallel ACSet transformations}
\usage{
equalizer(f, g)
}
\arguments{
\item{f}{ACSetTransformation A → B}

\item{g}{ACSetTransformation A → B}
}
\value{
List with \code{equalizer} (ACSet) and \code{incl} (inclusion morphism)
}
\description{
Given f, g: A → B, compute the subobject of A where f and g agree.
}
