% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fincat.R
\name{FinFunctor}
\alias{FinFunctor}
\title{Functor between finite categories (schema morphism)}
\usage{
FinFunctor(ob_map = list(), hom_map = list(), dom = FinCat(), codom = FinCat())
}
\arguments{
\item{ob_map}{Named list mapping domain objects to codomain objects}

\item{hom_map}{Named list mapping domain morphisms to codomain morphisms}

\item{dom}{Domain \link{FinCat}}

\item{codom}{Codomain \link{FinCat}}
}
\description{
Functor between finite categories (schema morphism)
}
