% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/homomorphism.R
\name{is_homomorphic}
\alias{is_homomorphic}
\title{Check if a homomorphism exists}
\usage{
is_homomorphic(pattern, target, monic = FALSE)
}
\arguments{
\item{pattern}{Source ACSet (typically small)}

\item{target}{Target ACSet (typically larger)}

\item{monic}{Logical; if TRUE, require injective components (monomorphism)}
}
\value{
Logical
}
\description{
Check if a homomorphism exists
}
\examples{
e <- path_graph(2)
tri <- cycle_graph(3)
is_homomorphic(e, tri) # TRUE
}
