% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fincat.R
\name{is_natural}
\alias{is_natural}
\title{Check naturality of an ACSet transformation}
\usage{
is_natural(alpha)
}
\arguments{
\item{alpha}{An \link{ACSetTransformation}}
}
\description{
Check naturality of an ACSet transformation
}
\examples{
g <- path_graph(2)
h <- path_graph(3)
alpha <- ACSetTransformation(
  components = list(V = c(1L, 2L), E = 1L),
  dom_acset = g, codom_acset = h
)
is_natural(alpha) # TRUE
}
