% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphviz.R
\name{petri_to_dot}
\alias{petri_to_dot}
\title{Render a Petri net as a bipartite DOT graph}
\usage{
petri_to_dot(pn)
}
\arguments{
\item{pn}{A Petri net ACSet}
}
\description{
Species are circles, transitions are boxes. Input/output arcs
connect them. Uses DiagrammeR for rendering.
}
