% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uwd.R
\name{relation}
\alias{relation}
\title{Create a wiring diagram using relation-style DSL}
\usage{
relation(..., .boxes = list())
}
\arguments{
\item{...}{Named outer port variables (junction names exposed to outside)}

\item{.boxes}{A list of box specifications, each a named list with
\code{name} and junction variable references}
}
\value{
A UWD ACSet
}
\description{
Create a wiring diagram using relation-style DSL
}
