% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rewriting.R
\name{rewrite}
\alias{rewrite}
\title{Apply a rule to a graph, finding the first valid match}
\usage{
rewrite(rule, graph)
}
\arguments{
\item{rule}{A Rule object}

\item{graph}{Target ACSet}
}
\value{
The rewritten ACSet, or NULL if no match found
}
\description{
Apply a rule to a graph, finding the first valid match
}
