% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphviz.R
\name{to_graphviz}
\alias{to_graphviz}
\title{Render an ACSet as an interactive diagram via DiagrammeR}
\usage{
to_graphviz(x, ...)
}
\arguments{
\item{x}{An ACSet (graph, Petri net, UWD, etc.)}

\item{...}{Passed to \code{to_dot()}}
}
\value{
A DiagrammeR \code{htmlwidget} (displays in RStudio Viewer or notebook)
}
\description{
Render an ACSet as an interactive diagram via DiagrammeR
}
\examples{
g <- path_graph(3)
\donttest{
to_graphviz(g)
}
}
