% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/typed_acsets.R
\name{typed_acset}
\alias{typed_acset}
\title{Create a typed ACSet}
\usage{
typed_acset(acset, type_system, ...)
}
\arguments{
\item{acset}{The concrete ACSet}

\item{type_system}{The type system ACSet}

\item{...}{Named integer vectors: component mappings for the typing morphism}
}
\value{
TypedACSet
}
\description{
Create a typed ACSet
}
\examples{
T <- Graph()
add_vertices(T, 2)
add_edge(T, 1, 2)
g <- Graph()
add_vertices(g, 3)
add_edge(g, 1, 3)
add_edge(g, 2, 3)
tg <- typed_acset(g, T, V = c(1L, 1L, 2L), E = c(1L, 1L))
flatten_typed(tg)
}
