% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/typed_acsets.R
\name{typed_coproduct}
\alias{typed_coproduct}
\title{Typed coproduct of typed ACSets}
\usage{
typed_coproduct(tacs1, tacs2)
}
\arguments{
\item{tacs1}{TypedACSet}

\item{tacs2}{TypedACSet}
}
\value{
TypedACSet
}
\description{
Given typed ACSets phi_1: X_1 -> T and phi_2: X_2 -> T, compute their
coproduct via pushout of the initial object. In practice this is the
disjoint union with combined typing.
}
