% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/typed_acsets.R
\name{TypedACSet}
\alias{TypedACSet}
\title{Typed ACSet: an ACSet with a typing morphism to a type system}
\usage{
TypedACSet(
  acset = acsets::ACSet(),
  type_system = acsets::ACSet(),
  typing = ACSetTransformation()
)
}
\arguments{
\item{acset}{The ACSet instance}

\item{type_system}{The type system ACSet}

\item{typing}{An \link{ACSetTransformation} from acset to type_system}
}
\description{
A TypedACSet pairs an ACSet X with a morphism phi: X -> T where T is the
"type system" (another ACSet on the same schema). Every element in X is
assigned a type by phi.
}
