% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphviz.R
\name{uwd_to_dot}
\alias{uwd_to_dot}
\title{Render a UWD as a DOT diagram}
\usage{
uwd_to_dot(w)
}
\arguments{
\item{w}{A UWD ACSet}
}
\description{
Boxes are rectangles, junctions are small circles,
outer ports are shown at the boundary.
}
\examples{
w <- uwd(c("s", "r"), c("s", "i"), c("i", "r"))
cat(uwd_to_dot(w))
}
